+++ common/src/main/java/dev/terminalmc/commandkeys/util/KeybindUtil.java
@@ -116,7 +116,7 @@ public class KeybindUtil {
                 
                 if (send) {
                     if (first) {
-                        ratelimited = !macro.ignoreRatelimit && !canTrigger(key);
+                        ratelimited = macro.useRatelimitStatus && !canTrigger(key);
                         first = false;
                     }
                     // Always allow repeat-stop